refactor: 잘못수정한 폴더 이름 원복
