+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.integration;
+package kitchenpos.ui;
 
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.domain.MenuGroup;
