+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.integration;
+package kitchenpos.ui;
 
 import kitchenpos.application.MenuService;
 import kitchenpos.domain.Menu;
